Split action_payloads into smaller modules to fix module_max_lines lint
Split the action_payloads module into four focused submodules:
- color.rs: Color and Rgb8 types
- float.rs: Float utilities, Points, UnitF32, Opacity, Degrees
- geometry.rs: Point, Position, Dimensions, Size
- stroke.rs: StrokeWidth, Rotation

This resolves the module_max_lines lint violation that counted both
action_payloads.rs and action_payloads_tests.rs as a single module.
