+++ crates/gauss-core/src/model/action_payloads/float.rs
@@ -1,0 +1,102 @@
+//! Float normalization and validation utilities.
+
+/// Normalize a float for consistent hashing and equality.
+///
+/// Maps -0.0 to 0.0. NaN values should be rejected before calling this.
+#[inline]
+pub const fn normalize_float(value: f32) -> f32 {
+    if value == 0.0 { 0.0 } else { value }
+}
+
+/// Helper to validate and normalize a pair of finite f32 values.
+#[inline]
+pub(crate) const fn new_pair(x: f32, y: f32) -> Option<(f32, f32)> {
+    if x.is_finite() && y.is_finite() {
+        Some((normalize_float(x), normalize_float(y)))
+    } else {
+        None
+    }
+}
+
+/// Generic length unit in document points.
+#[derive(Clone, Copy, Debug, Default, PartialEq, PartialOrd)]
+pub struct Points(pub f32);
+
+impl Points {
+    /// Return the inner value.
+    #[must_use]
+    pub const fn value(&self) -> f32 {
+        self.0
+    }
+}
+
+/// Unit float in the range [0.0, 1.0].
+#[derive(Clone, Copy, Debug, Default, PartialEq, PartialOrd)]
+pub struct UnitF32(f32);
+
+impl UnitF32 {
+    /// Return the inner value.
+    #[must_use]
+    pub const fn value(&self) -> f32 {
+        self.0
+    }
+}
+
+impl TryFrom<f32> for UnitF32 {
+    type Error = &'static str;
+
+    fn try_from(v: f32) -> Result<Self, Self::Error> {
+        if v.is_finite() && (0.0..=1.0).contains(&v) {
+            Ok(Self(v))
+        } else {
+            Err("value out of range, expected 0.0..=1.0")
+        }
+    }
+}
+
+/// Opacity value (0.0..=1.0).
+#[derive(Clone, Copy, Debug, Default, PartialEq, PartialOrd)]
+pub struct Opacity(f32);
+
+impl Opacity {
+    /// Construct a new opacity.
+    ///
+    /// Returns `None` if the value is not in the range [0.0, 1.0].
+    #[must_use]
+    pub const fn new(value: UnitF32) -> Option<Self> {
+        // UnitF32 already validates the range, but we double-check for const fn
+        if !value.0.is_finite() {
+            return None;
+        }
+        if value.0 < 0.0 || value.0 > 1.0 {
+            return None;
+        }
+        Some(Self(normalize_float(value.0)))
+    }
+
+    /// Return the opacity value.
+    #[must_use]
+    pub const fn value(&self) -> f32 {
+        self.0
+    }
+}
+
+impl std::hash::Hash for Opacity {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.0.to_bits().hash(state);
+    }
+}
+
+impl Eq for Opacity {}
+
+/// Angle in degrees.
+#[derive(Clone, Copy, Debug, Default, PartialEq, PartialOrd)]
+pub struct Degrees(pub f32);
+
+impl Degrees {
+    /// Return the inner value.
+    #[must_use]
+    pub const fn value(&self) -> f32 {
+        self.0
+    }
+}
